requirements.txt
Flask==2.0.1
requests==2.26.0
numpy==1.21.4
pandas==1.3.4
mysql-connector-python==8.0.27
